Refactor alphanumeric into its own module
